Created a basic Geography Quiz
Users answer the related questions and at the end, they receive a results page that gives them the score. I added separate arrays that also tallied provided the question, number, and answers, both correct and incorrect. Based on the answers correct, you got a message that went with the score.